[FIX] separate AssetId definition to be more importable
