+++ pandora-common/src/assets/appearance.ts
@@ -4,13 +4,14 @@ import type { ActionRoomContext } from '../chatroom';
 import { Assert } from '../utility';
 import { EvalItemPath } from './appearanceHelpers';
 import type { ItemPath, RoomActionTargetCharacter } from './appearanceTypes';
-import { AppearanceItems } from './appearanceValidation';
-import { AssetManager } from './assetManager';
-import { AssetId, WearableAssetType } from './definitions';
+import type { AppearanceItems } from './appearanceValidation';
+import type { AssetManager } from './assetManager';
+import type { WearableAssetType } from './definitions';
+import type { AssetId } from './base';
 import type { BoneState, CharacterView } from './graphics';
-import { Item } from './item';
-import { AppearanceArmPose, AppearanceBundle, AppearancePose, AssetFrameworkCharacterState, SafemodeData } from './state/characterState';
-import { GameLogicCharacter } from '../gameLogic/character/character';
+import type { Item } from './item';
+import type { AppearanceArmPose, AppearanceBundle, AppearancePose, AssetFrameworkCharacterState, SafemodeData } from './state/characterState';
+import type { GameLogicCharacter } from '../gameLogic/character/character';
 
 export const BONE_MIN = -180;
 export const BONE_MAX = 180;
